+++ app/hooks/queries.ts
@@ -222,12 +222,13 @@ export const useIntegrations = (
 
 export const useRunner = (
   variables: RunnerQueryVariables,
-  skip?: boolean
+  { pollInterval, skip }: { pollInterval: number; skip?: boolean }
 ): QueryResult<RunnerData, RunnerQueryVariables> => {
   return useQuery<RunnerData, RunnerQueryVariables>(runnerQuery, {
     fetchPolicy,
     nextFetchPolicy,
     onError,
+    pollInterval,
     skip: (!variables.run_id && !variables.test_id) || skip,
     variables,
   });
